released ver 6.1.7

+++ src/jp/or/med/orca/qkan/affair/qs/qs001/QS001.java
@@ -1447,20 +1447,26 @@ public class QS001 extends QS001Event {
                 
                 // 2013/10/12 [Shinobu_Hitaka] add - begin
                 // 同日同サービスがリストの下にある時、SERVICE_IDをクリアして順位を保持する（登録時にIDを振りなおさせる為）
-                VRList selServices = getSelectedServiceListBox().getSchedule();
-                int selServicesIndex = getSelectedServiceListBox().getSelectedIndex();
-                Date srcDate = ACCastUtilities.toDate(service.get("SERVICE_DATE"));
-                String srcSystemServiceKindDetail = ACCastUtilities.toString(service.get("SYSTEM_SERVICE_KIND_DETAIL"), "");
-                // 配列の後方を検索
-                for (int i = selServicesIndex+1; i < selServices.size(); i++) {
-                    VRMap destService = (VRMap) selServices.get(i);
-                    Date destDate = ACCastUtilities.toDate(destService.get("SERVICE_DATE"));
-                    String destSystemServiceKindDetail = ACCastUtilities.toString(destService.get("SYSTEM_SERVICE_KIND_DETAIL"), "");
-                    // 日付とサービス種類コードが同じ場合
-                    if (srcDate.equals(destDate) && srcSystemServiceKindDetail.equals(destSystemServiceKindDetail)){
-                    	// サービスIDを削除
-                    	destService.remove("SERVICE_ID");
-                    }
+                if (getSelectedServiceListBox() != null) {
+	                VRList selServices = getSelectedServiceListBox().getSchedule();
+	                if (selectedIndex >= 0){
+	                	Object selDate = service.get("SERVICE_DATE");
+	                	if (selDate != null) {
+			                Date srcDate = ACCastUtilities.toDate(selDate);
+			                String srcSystemServiceKindDetail = ACCastUtilities.toString(service.get("SYSTEM_SERVICE_KIND_DETAIL"), "");
+			                // 配列の後方を検索
+			                for (int i = selectedIndex+1; i < selServices.size(); i++) {
+			                    VRMap destService = (VRMap) selServices.get(i);
+			                    Date destDate = ACCastUtilities.toDate(destService.get("SERVICE_DATE"));
+			                    String destSystemServiceKindDetail = ACCastUtilities.toString(destService.get("SYSTEM_SERVICE_KIND_DETAIL"), "");
+			                    // 日付とサービス種類コードが同じ場合
+			                    if (srcDate.equals(destDate) && srcSystemServiceKindDetail.equals(destSystemServiceKindDetail)){
+			                    	// サービスIDを削除
+			                    	destService.remove("SERVICE_ID");
+			                    }
+			                }
+	                	}
+	                }
                 }
                 // 2013/10/12 [Shinobu_Hitaka] add - end
